pydantic-ai-rag/requirements.txt
asyncpg>=0.28.0
python-dotenv>=1.0.0
openai>=1.0.0
pydantic-ai>=0.0.14